factor common stuff out of handlers
